Ok, I have a handle on this now. In order to have LFS work on forks, you don't need to mirror or do any kind of special cloning like that. Instead just clone it normally, create a repo, and push to that origin branch.
https://docs.github.com/en/billing/managing-billing-for-git-large-file-storage/about-billing-for-git-large-file-storage

The free version only has 1gb of storage though, and I am guessing the bandwidth refers to the uploads. It really isn't much.

But I am happy that unlike with the `migration import`, it didn't just swap all the files with pointers. I think I was using `--fixup` for this.

That handles this. I'll want to look into how to dockerize the application next. Or at least, the server.

Let me just try cloning it locally to make sure that the model gets transfered. It is right there in the online repo, but I do not want any surprises when I download it.

...Yeah, it works great. Just a regular clone is enough to pull the LFS objects. I thought that I might need some separate commands, but it doesn't seem to be the case.