https://stackoverflow.com/a/66644457/4364027
This solution is for Poetry, but I think that kicking PyTorch out from the Pipfile and installing it and the dependencies related to it using regular pip is the way to go today for any package manager.

This solution has my vote.

Lest I forget, when I tried running the Monadical project for the first time with the 1.13.0 version, it actually couldn't install the dependency properly. So there is no point in having it there anyway if it is just going to break a few months down the road.