``` poetry source add -p explicit pytorch https://download.pytorch.org/whl/cu117/ poetry add --source pytorch torch torchvision ```
Peotry doesn't give progress reporting during dependecy resolution either. According to some of the Github issues, it needs to download the wheels in order to resolve them, and they'd probably come to something like 2gb. Maybe if I waited half an hour, the dependencies would resolve, but I'd stick with my earlier suggestion. Even if running the above could technically work, I am not the type to trust the computer to do the right thing without any feedback for such a long time.

I just had to check out Poetry to see if it does a better job than the rest, and it doesn't seem like it.