During the interview the copy image URL didn't work as expected. It seems Blazor WASM is more permissive when it comes to fetching than JS.
I am not going to be working on this project anymore as it turns out the interviews aren't programming intense, but when it comes to containers let me write one last thought.

Maybe the reason why I am just getting connection closed kind of errors when running Azure containers, is due to running the frontend in production mode. It might be worrth trying it out in dev mode.

Apart from that, the only other idea is to make 2 minimalist servers and experiment on those.

Honestly, working on this was fun, as I learned a bunch about Python package management and docker, as well as deployment on Azure. This will all be useful knowledge in the future. As a pro, it is something that I should be familiar either way.